Huge EDDConfig update
Remove the useless EDDConfig.MapColoursClass.
Added an initializing prop to Settings and EDDiscoveryForm to prevent
	recursion during init. Will be removed from Settings once
	binding is done, TODO: ?not sure about EDDiscoveryForm?
Added EDDConfig.HomeSystem property. (TODO: need to write changes)
Added proper lazy loading with thread safety for EDDConfig (note that
	the safety only applies to init, which will only occur once...)
Cleaned up some EDCommander.Load issues when not using -appfolder opt.
EDDConfig Properties can now all be accessed statically.
Move EDDConfig from EDDiscovery2 namespace to EDDiscovery namespace.

+++ EDDiscovery/EliteDangerous/EDCommander.cs
@@ -303,29 +303,31 @@ namespace EDDiscovery2
                 }
             }
 
-            JObject jo = new JObject();
-            foreach (EDCommander cmdr in _commandersDict.Values)
+            if (!string.IsNullOrWhiteSpace(EDDConfig.Options.AppFolder))
             {
-                jo[cmdr.Name] = new JObject(new
+                JObject jo = new JObject();
+                foreach (EDCommander cmdr in _commandersDict.Values)
                 {
-                    NetLogDir = cmdr.NetLogDir,
-                    JournalDir = cmdr.JournalDir
-                });
-            }
+                    jo[cmdr.Name] = new JObject(
+                        new JProperty("NetLogDir", cmdr.NetLogDir),
+                        new JProperty("JournalDir", cmdr.JournalDir)
+                        );
+                }
 
-            using (Stream stream = File.OpenWrite(Path.Combine(EDDConfig.Options.AppFolder, "CommanderPaths.json.tmp")))
-            {
-                using (StreamWriter writer = new StreamWriter(stream))
+                using (Stream stream = File.OpenWrite(Path.Combine(EDDConfig.Options.AppFolder, "CommanderPaths.json.tmp")))
                 {
-                    using (JsonTextWriter jwriter = new JsonTextWriter(writer))
+                    using (StreamWriter writer = new StreamWriter(stream))
                     {
-                        jo.WriteTo(jwriter);
+                        using (JsonTextWriter jwriter = new JsonTextWriter(writer))
+                        {
+                            jo.WriteTo(jwriter);
+                        }
                     }
                 }
-            }
 
-            File.Delete(Path.Combine(EDDConfig.Options.AppFolder, "CommanderPaths.json"));
-            File.Move(Path.Combine(EDDConfig.Options.AppFolder, "CommanderPaths.json.tmp"), Path.Combine(EDDConfig.Options.AppFolder, "CommanderPaths.json"));
+                File.Delete(Path.Combine(EDDConfig.Options.AppFolder, "CommanderPaths.json"));
+                File.Move(Path.Combine(EDDConfig.Options.AppFolder, "CommanderPaths.json.tmp"), Path.Combine(EDDConfig.Options.AppFolder, "CommanderPaths.json"));
+            }
         }
 
         /// <summary>
@@ -426,7 +428,7 @@ namespace EDDiscovery2
                 }
             }
 
-            if (File.Exists(Path.Combine(EDDConfig.Options.AppFolder, "CommanderPaths.json")))
+            if (!string.IsNullOrWhiteSpace(EDDConfig.Options.AppFolder) && File.Exists(Path.Combine(EDDConfig.Options.AppFolder, "CommanderPaths.json")))
             {
                 JObject jo;
